Clean tests + Fix newline and UTF8 support
