+++ composer.json
@@ -7,7 +7,7 @@
     },
     "require-dev": {
         "laravel/framework": "^5.1.0",
-        "phpunit/phpunit": "~5.7"
+        "phpunit/phpunit": "^7.0"
     },
     "license": "MIT",
     "authors": [
@@ -24,6 +24,11 @@
     "autoload": {
         "psr-4": {
             "Alsofronie\\Uuid\\": "src/"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Tests\\": "tests/"
         }
     },
     "extra": {
