+++ phpunit.xml
@@ -2,7 +2,7 @@
 <phpunit
         backupGlobals="false"
         backupStaticAttributes="false"
-        bootstrap="tests/bootstrap.php"
+        bootstrap="vendor/autoload.php"
         convertErrorsToExceptions="true"
         convertNoticesToExceptions="true"
         convertWarningsToExceptions="true"
@@ -11,7 +11,6 @@
         stopOnFailure="false"
         stopOnIncomplete="false"
         stopOnSkipped="false"
-        syntaxCheck="false"
         verbose="true"
         colors="true"
         >
@@ -22,7 +21,7 @@
     </filter>
     <testsuites>
         <testsuite name="default">
-            <directory>./tests/</directory>
+            <directory>./tests</directory>
         </testsuite>
     </testsuites>
 </phpunit>
