+++ tests/DatabaseIntegrationTestConnectionResolver.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace Tests;
+
+class DatabaseIntegrationTestConnectionResolver implements \Illuminate\Database\ConnectionResolverInterface
+{
+    protected $connection;
+
+    public function connection($name = null)
+    {
+        if (isset($this->connection)) {
+            return $this->connection;
+        }
+
+        return $this->connection = new \Illuminate\Database\SQLiteConnection(new \PDO('sqlite::memory:'));
+    }
+
+    public function getDefaultConnection()
+    {
+        return 'default';
+    }
+
+    public function setDefaultConnection($name)
+    {
+        //
+    }
+}
